+++ packages/vue-script-setup-converter/src/lib/helpers/node.ts
@@ -1,4 +1,3 @@
-// TODO: Move to helpers/node.ts
 import { SyntaxKind, Node, PropertyAssignment, CallExpression } from "ts-morph";
 
 export const getNodeByKind = (
